requirements.txt
torch>=2.0
torchvision
numpy
matplotlib
tensorboard
opencv-python
Pillow
